Splitting the GaussianWavefront into GaussianWavefront and GaussianPropagator
