+++ src/propagators.py
@@ -232,3 +232,193 @@ class FresnelProp(eqx.Module):
         wavefront_out = norm_coeff * t2
 
         return wavefront_out
+
+
+class GaussianPropagator(Propagator):
+
+
+    # Constants
+    INDEX_GENERATOR = numpy.array([1, 2])
+
+
+    def planar_to_planar(self: FresnelWavefront, distance: float) -> None:
+        """
+        Modifies the state of the wavefront by propagating a planar 
+        wavefront to a planar wavefront. 
+
+        Parameters
+        ----------
+        distance : float
+            The distance of the propagation in metres.
+        """
+        wavefront = self.amplitude * \
+            numpy.exp(1j * self.phase)
+
+        new_wavefront = numpy.fft.ifft2(
+            self.transfer_function(distance) * \
+            numpy.fft.fft2(wavefront))
+
+        amplitude = numpy.abs(new_wavefront)
+        phase = self.calculate_phase(new_wavefront)
+        return self.update_phasor(amplitude, phase)        
+
+
+    def waist_to_spherical(self: FresnelWavefront, 
+            distance: float) -> None:
+        """
+        Modifies the state of the wavefront by propagating it from 
+        the waist of the gaussian beam to a spherical wavefront. 
+
+        Parameters
+        ----------
+        distance : float 
+            The distance of the propagation in metres.
+        """
+        coefficient = 1 / 1j / self.wavel / distance
+
+        wavefront = self.amplitude * numpy.exp(1j * self.phase)
+
+        # TODO: Check that these are the correct algorithms to 
+        # use and ask if we are going to actually require the 
+        # reverse direction
+        fourier_transform = jax.lax.cond(numpy.sign(distance) > 0, 
+            lambda wavefront, distance: \
+                quadratic_phase_factor(distance) * \
+                numpy.fft.fft2(wavefront), 
+            lambda wavefront, distance: \
+                quadratic_phase_factor(distance) * \
+                numpy.fft.ifft2(wavefront),
+            wavefront, distance)
+
+        new_wavefront = coefficient * fourier_transform
+        phase = self.calculate_phase(new_wavefront)
+        amplitude = jax.numpy.abs(new_wavefront)
+        self.update_phasor(amplitude, phase)
+
+        # TODO: Confirm that I need to update the pixel scale for 
+        # this transformation.
+
+
+    def spherical_to_waist(self: FresnelWavefront, 
+            distance: float) -> None:
+        """
+        Modifies the state of the wavefront by propagating it from 
+        a spherical wavefront to the waist of the Gaussian beam. 
+
+        Parameters
+        ----------
+        distance : float
+            The distance of propagation in metres
+        """
+        coefficient = 1 / 1j / self.wavel / distance * \
+            quadratic_phase_factor(distance)
+
+        wavefront = self.amplitude * numpy.exp(1j * self.phase)
+
+        # TODO: Check that these are the correct algorithms to 
+        # use and ask if we are going to actually require the 
+        # reverse direction
+        fourier_transform = jax.lax.cond(numpy.sign(distance) > 0, 
+            lambda wavefront: numpy.fft.fft2(wavefront), 
+            lambda wavefront: numpy.fft.ifft2(wavefront),
+            wavefront)
+
+        new_wavefront = coefficient * fourier_transform
+        phase = self.calculate_phase(new_wavefront)
+        amplitude = jax.numpy.abs(new_wavefront)
+        self.update_phasor(amplitude, phase)
+
+        # TODO: Confirm that I need to update the pixel scale for 
+        # this transformation.
+
+
+    def outside_to_outside(self: FresnelWavefront, distance: float) -> None:
+        """
+        Propagation from outside the Rayleigh range to another 
+        position outside the Rayleigh range. 
+
+        Parameters
+        ----------
+        distance : float
+            The distance to propagate in metres.
+        """
+        self.waist_to_spherical(self.position + distance - \
+            self.location_of_waist()) * self.waist_to_spherical(
+            self.location_of_waist() - self.position)
+
+
+    def outside_to_inside(self: FresnelWavefront, distance: float) -> None:
+        """
+        Propagation from outside the Rayleigh range to inside the 
+        rayleigh range. 
+
+        Parameters
+        ----------
+        distance : float
+            The distance to propagate in metres.
+        """
+        self.planar_to_planar(self.position + distance - \
+            self.location_of_waist()) * self.spherical_to_waist(
+            self.location_of_waist() - self.position)
+
+
+    def inside_to_inside(self: FresnelWavefront, distance: float) -> None:
+        """
+        Propagation from inside the Rayleigh range to inside the 
+        rayleigh range. 
+
+        Parameters
+        ----------
+        distance : float
+            The distance to propagate in metres.
+        """
+        # Consider removing after checking the Jaxpr for speed. 
+        # This is just an alias for another function. 
+        self.planar_to_planar(distance)
+
+
+    def inside_to_outside(self: FresnelWavefront, distance: float) -> None:
+        """
+        Propagation from inside the Rayleigh range to outside the 
+        rayleigh range. 
+
+        Parameters
+        ----------
+        distance : float
+            The distance to propagate in metres.
+        """
+        self.waist_to_spherical(self.position + distance - \
+            self.location_of_waist()) * planar_to_planar(
+            self.location_of_waist() - self.position)
+
+
+    def propagate(self: FresnelWavefront, distance: float) -> None:
+        """
+        Propagates the wavefront approximated by a Gaussian beam 
+        the amount specified by distance. Note that distance can 
+        be negative.
+
+        Parameters 
+        ----------
+        distance : float
+            The distance to propagate in metres.
+        """
+        # This works by considering the current position and distnace 
+        # as a boolean array. The INDEX_GENERATOR converts this to 
+        # and index according to the following table.
+        #
+        # sum((0, 0) * (1, 2)) == 0
+        # sum((1, 0) * (1, 2)) == 1
+        # sum((0, 1) * (1, 2)) == 2
+        # sum((1, 1) * (1, 2)) == 3
+        #
+        # TODO: Test if a simple lookup is faster. 
+        decision_vector = self.is_inside([0., distance])
+        decision_index = numpy.sum(self.INDEX_GENERATOR * descision_vector)
+ 
+        # Enters the correct function differentiably depending on 
+        # the descision.
+        jax.lax.switch(decision_index, 
+            [self.inside_to_inside, self.inside_to_outside,
+            self.outside_to_inside, self.outside_to_outside],
+            distance)
