+++ src/wavefronts.py
@@ -668,23 +668,23 @@ class GaussianWavefront(dLux.PhysicalWavefront):
     radius of the beam, the phase radius, the phase factor and the 
     Rayleigh distance. Propagation is based on two different regimes 
     for a total of four different opertations. 
+    
+    Attributes
+    ----------
+    position : float
+        The position of the wavefront in the optical system assumed 
+        to be in meters.
+    beam_radius : float
+        The radius of the beam assumed to be in metres. 
+    phase_radius : float
+        The phase radius of the gaussian beam assumed to be unitless. 
     """
-    # Dunder attributes
-    # TODO: Test and profile a working slots implementation.
-    # __slots__ = ("INDEX_GENERATOR", "position", "wavel")
-
-    # Constants
-    INDEX_GENERATOR = numpy.array([1, 2])
-
-    # Variables
     position : float 
-    beam_radius : Array
-    phase_radius : 
+    beam_radius : float
+    phase_radius : float
 
-
-    def __init__(self: FresnelWavefront, beam_radius: Array, 
-            wavel: float, offset: Array, amplitude: Array, 
-            phase: Array, phase_radius: float, 
+    def __init__(self: FresnelWavefront, beam_radius: float, 
+            wavel: float, phase_radius: float, 
             position: float=0.0) -> FresnelWavefront:
         """
         Creates a wavefront with an empty amplitude and phase 
@@ -704,18 +704,17 @@ class GaussianWavefront(dLux.PhysicalWavefront):
         phase : Array
             An array containing the electric field phase over the 
             wavefront. Assumed to be square.
-        
         """
-        super().__init__(wavel, offset)
+        # TODO: Is offset required in this wavefront
+        super().__init__(wavel)
         self.beam_radius = beam_radius
-        self.amplitude = amplitude
-        self.phase = phase
+        self.phase_radius = phase_radius
+        self.amplitude = None
+        self.phase = None
         self.position = position
-        # TODO: Determine if I need to implement some method of 
-        # best fit for the phase_radius and beam_radius. 
 
 
-    def rayleigh_distance(self: FresnelWavefront) -> None:
+    def rayleigh_distance(self: FresnelWavefront) -> float:
         """
         Calculates the rayleigh distance of the Gaussian beam.
         
@@ -725,97 +724,6 @@ class GaussianWavefront(dLux.PhysicalWavefront):
             The Rayleigh distance of the wavefront in metres.
         """
         return numpy.pi * self.beam_radius ** 2 / self.wavel
-
-
-    def planar_to_planar(self: FresnelWavefront, distance: float) -> None:
-        """
-        Modifies the state of the wavefront by propagating a planar 
-        wavefront to a planar wavefront. 
-
-        Parameters
-        ----------
-        distance : float
-            The distance of the propagation in metres.
-        """
-        wavefront = self.amplitude * \
-            numpy.exp(1j * self.phase)
-
-        new_wavefront = numpy.fft.ifft2(
-            self.transfer_function(distance) * \
-            numpy.fft.fft2(wavefront))
-
-        amplitude = numpy.abs(new_wavefront)
-        phase = self.calculate_phase(new_wavefront)
-        return self.update_phasor(amplitude, phase)        
-
-
-    def waist_to_spherical(self: FresnelWavefront, 
-            distance: float) -> None:
-        """
-        Modifies the state of the wavefront by propagating it from 
-        the waist of the gaussian beam to a spherical wavefront. 
-
-        Parameters
-        ----------
-        distance : float 
-            The distance of the propagation in metres.
-        """
-        coefficient = 1 / 1j / self.wavel / distance
-
-        wavefront = self.amplitude * numpy.exp(1j * self.phase)
-
-        # TODO: Check that these are the correct algorithms to 
-        # use and ask if we are going to actually require the 
-        # reverse direction
-        fourier_transform = jax.lax.cond(numpy.sign(distance) > 0, 
-            lambda wavefront, distance: \
-                quadratic_phase_factor(distance) * \
-                numpy.fft.fft2(wavefront), 
-            lambda wavefront, distance: \
-                quadratic_phase_factor(distance) * \
-                numpy.fft.ifft2(wavefront),
-            wavefront, distance)
-
-        new_wavefront = coefficient * fourier_transform
-        phase = self.calculate_phase(new_wavefront)
-        amplitude = jax.numpy.abs(new_wavefront)
-        self.update_phasor(amplitude, phase)
-
-        # TODO: Confirm that I need to update the pixel scale for 
-        # this transformation.
-
-
-    def spherical_to_waist(self: FresnelWavefront, 
-            distance: float) -> None:
-        """
-        Modifies the state of the wavefront by propagating it from 
-        a spherical wavefront to the waist of the Gaussian beam. 
-
-        Parameters
-        ----------
-        distance : float
-            The distance of propagation in metres
-        """
-        coefficient = 1 / 1j / self.wavel / distance * \
-            quadratic_phase_factor(distance)
-
-        wavefront = self.amplitude * numpy.exp(1j * self.phase)
-
-        # TODO: Check that these are the correct algorithms to 
-        # use and ask if we are going to actually require the 
-        # reverse direction
-        fourier_transform = jax.lax.cond(numpy.sign(distance) > 0, 
-            lambda wavefront: numpy.fft.fft2(wavefront), 
-            lambda wavefront: numpy.fft.ifft2(wavefront),
-            wavefront)
-
-        new_wavefront = coefficient * fourier_transform
-        phase = self.calculate_phase(new_wavefront)
-        amplitude = jax.numpy.abs(new_wavefront)
-        self.update_phasor(amplitude, phase)
-
-        # TODO: Confirm that I need to update the pixel scale for 
-        # this transformation.
 
 
     def transfer_function(self: FresnelWavefront, distance: float) -> Array:
@@ -926,69 +834,10 @@ class GaussianWavefront(dLux.PhysicalWavefront):
         """
         number_of_pixels = self.amplitude.shape[0]
         new_pixel_scale = self.wavel * numpy.abs(position) / \
-            number_of_pixels / self.pixel_scale     
+            number_of_pixels / self.pixel_scale  
+        return equinox.tree_at(lambda wavefront : self.pixel_scale,
+            self, new_pixel_scale)   
         
-
-
-    def outside_to_outside(self: FresnelWavefront, distance: float) -> None:
-        """
-        Propagation from outside the Rayleigh range to another 
-        position outside the Rayleigh range. 
-
-        Parameters
-        ----------
-        distance : float
-            The distance to propagate in metres.
-        """
-        self.waist_to_spherical(self.position + distance - \
-            self.location_of_waist()) * self.waist_to_spherical(
-            self.location_of_waist() - self.position)
-
-
-    def outside_to_inside(self: FresnelWavefront, distance: float) -> None:
-        """
-        Propagation from outside the Rayleigh range to inside the 
-        rayleigh range. 
-
-        Parameters
-        ----------
-        distance : float
-            The distance to propagate in metres.
-        """
-        self.planar_to_planar(self.position + distance - \
-            self.location_of_waist()) * self.spherical_to_waist(
-            self.location_of_waist() - self.position)
-
-
-    def inside_to_inside(self: FresnelWavefront, distance: float) -> None:
-        """
-        Propagation from inside the Rayleigh range to inside the 
-        rayleigh range. 
-
-        Parameters
-        ----------
-        distance : float
-            The distance to propagate in metres.
-        """
-        # Consider removing after checking the Jaxpr for speed. 
-        # This is just an alias for another function. 
-        self.planar_to_planar(distance)
-
-
-    def inside_to_outside(self: FresnelWavefront, distance: float) -> None:
-        """
-        Propagation from inside the Rayleigh range to outside the 
-        rayleigh range. 
-
-        Parameters
-        ----------
-        distance : float
-            The distance to propagate in metres.
-        """
-        self.waist_to_spherical(self.position + distance - \
-            self.location_of_waist()) * planar_to_planar(
-            self.location_of_waist() - self.position)
-
 
     def is_inside(self: FresnelWavefront, distance: float) -> Boolean:
         """ 
@@ -1007,37 +856,7 @@ class GaussianWavefront(dLux.PhysicalWavefront):
             true if the point is within the rayleigh distance false 
             otherwise.
         """
-        return float(numpy.abs(self.position + distance - \
-            self.location_of_waist()) <= self.rayleigh_distance())
-        
-
-    def propagate(self: FresnelWavefront, distance: float) -> None:
-        """
-        Propagates the wavefront approximated by a Gaussian beam 
-        the amount specified by distance. Note that distance can 
-        be negative.
-
-        Parameters 
-        ----------
-        distance : float
-            The distance to propagate in metres.
-        """
-        # This works by considering the current position and distnace 
-        # as a boolean array. The INDEX_GENERATOR converts this to 
-        # and index according to the following table.
-        #
-        # sum((0, 0) * (1, 2)) == 0
-        # sum((1, 0) * (1, 2)) == 1
-        # sum((0, 1) * (1, 2)) == 2
-        # sum((1, 1) * (1, 2)) == 3
-        #
-        # TODO: Test if a simple lookup is faster. 
-        decision_vector = self.is_inside([0., distance])
-        decision_index = numpy.sum(self.INDEX_GENERATOR * descision_vector)
- 
-        # Enters the correct function differentiably depending on 
-        # the descision.
-        jax.lax.switch(decision_index, 
-            [self.inside_to_inside, self.inside_to_outside,
-            self.outside_to_inside, self.outside_to_outside],
-            distance) 
+        # TODO: Review the location of this function which increases
+        # coupling with the GaussianPropagator. 
+        return numpy.abs(self.position + distance - \
+            self.location_of_waist()) <= self.rayleigh_distance()
